Updated install.sh and removed emacs backups

+++ install.sh
@@ -3,16 +3,20 @@
 BASEDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 # bash
-ln -sv ${BASEDIR}/.bashrc ~/.bashrc
+if [ ! -L "${HOME}/.bashrc" ]; then
+    ln -sv ${BASEDIR}/.bashrc ~/.bashrc
+fi    
 
 # scripts
 # Create directory ~/.bin if it doesn't exit, then place all scripts there
-if [ ! -d "~/.bin" ]; then
+if [ ! -d "${HOME}/.bin" ]; then
     mkdir ~/.bin
 fi
 
 SCRIPTS="${BASEDIR}/bin/*"
 for file in ${SCRIPTS}
 do
-    ln -sv "${file}" "${HOME}/.bin/$(basename ${file})"
+    if [ ! -L "${HOME}/.bin/$(basename ${file})" ]; then
+      	ln -sv "${file}" "${HOME}/.bin/$(basename ${file})"
+    fi
 done
